+++ src/main/java/net/geforcemods/securitycraft/screen/BlockReinforcerScreen.java
@@ -1,9 +1,16 @@
 package net.geforcemods.securitycraft.screen;
 
+import java.io.IOException;
+import java.util.Arrays;
+
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.inventory.BlockReinforcerMenu;
+import net.geforcemods.securitycraft.network.server.SyncBlockReinforcer;
+import net.geforcemods.securitycraft.screen.components.CallbackCheckbox;
+import net.geforcemods.securitycraft.screen.components.StringHoverChecker;
 import net.geforcemods.securitycraft.util.GuiUtils;
 import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.inventory.Container;
@@ -12,20 +19,36 @@ import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 
 public class BlockReinforcerScreen extends GuiContainer {
-	private static final ResourceLocation TEXTURE = new ResourceLocation(SecurityCraft.MODID + ":textures/gui/container/universal_block_reinforcer.png");
-	private static final ResourceLocation TEXTURE_LVL1 = new ResourceLocation(SecurityCraft.MODID + ":textures/gui/container/universal_block_reinforcer_lvl1.png");
-	private final boolean isLvl1;
+	private static final ResourceLocation TEXTURE = new ResourceLocation(SecurityCraft.MODID, "textures/gui/container/universal_block_reinforcer.png");
+	private static final ResourceLocation TEXTURE_LVL1 = new ResourceLocation(SecurityCraft.MODID, "textures/gui/container/universal_block_reinforcer_lvl1.png");
+	private boolean isLvl1;
+	private CallbackCheckbox unreinforceCheckbox;
+	private StringHoverChecker checkboxHoverChecker;
 
 	public BlockReinforcerScreen(Container container, boolean isLvl1) {
 		super(container);
-
 		this.isLvl1 = isLvl1;
+		ySize = 186;
+	}
+
+	@Override
+	public void initGui() {
+		super.initGui();
+		unreinforceCheckbox = addButton(new CallbackCheckbox(0, guiLeft + 24, guiTop + 69, 20, 20, "", !((BlockReinforcerMenu) inventorySlots).isReinforcing, state -> {}, 0));
+
+		if (isLvl1)
+			unreinforceCheckbox.visible = false;
+
+		checkboxHoverChecker = new StringHoverChecker(unreinforceCheckbox, Arrays.asList(Utils.localize("gui.securitycraft:blockReinforcer.unreinforceCheckbox.not_checked").getFormattedText(), Utils.localize("gui.securitycraft:blockReinforcer.unreinforceCheckbox.checked").getFormattedText()));
 	}
 
 	@Override
 	public void drawScreen(int mouseX, int mouseY, float partialTicks) {
 		super.drawScreen(mouseX, mouseY, partialTicks);
 		renderHoveredToolTip(mouseX, mouseY);
+
+		if (checkboxHoverChecker.checkHover(mouseX, mouseY))
+			drawHoveringText(checkboxHoverChecker.getName(), mouseX, mouseY);
 	}
 
 	@Override
@@ -60,5 +83,17 @@ public class BlockReinforcerScreen extends GuiContainer {
 		GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
 		mc.getTextureManager().bindTexture(isLvl1 ? TEXTURE_LVL1 : TEXTURE);
 		drawTexturedModalRect(guiLeft, guiTop, 0, 0, xSize, ySize);
+	}
+
+	@Override
+	protected void actionPerformed(GuiButton button) throws IOException {
+		if (button == unreinforceCheckbox)
+			unreinforceCheckbox.onClick();
+	}
+
+	@Override
+	public void onGuiClosed() {
+		super.onGuiClosed();
+		SecurityCraft.network.sendToServer(new SyncBlockReinforcer(!unreinforceCheckbox.selected()));
 	}
 }