+++ src/main/java/net/geforcemods/securitycraft/screen/components/CallbackCheckbox.java
@@ -7,7 +7,7 @@ import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.MathHelper;
 
-public class CallbackCheckbox extends ClickButton {
+public class CallbackCheckbox extends ClickButton implements IToggleableButton {
 	private static final ResourceLocation TEXTURE = new ResourceLocation("securitycraft", "textures/gui/checkbox.png");
 	private boolean selected;
 	private final Consumer<Boolean> onChange;
@@ -28,13 +28,15 @@ public class CallbackCheckbox extends ClickButton {
 
 	@Override
 	public void drawButton(Minecraft minecraft, int mouseX, int mouseY, float partialTicks) {
-		minecraft.getTextureManager().bindTexture(TEXTURE);
-		GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-		GlStateManager.enableBlend();
-		GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
-		GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
-		drawScaledCustomSizeModalRect(x, y, 0, selected ? 20.0F : 0.0F, 20, 20, width, height, 32, 64);
-		minecraft.fontRenderer.drawString(displayString, x + (int) (width * 1.2F), y + (height - 8) / 2, textColor | MathHelper.ceil(255.0F) << 24);
+		if (visible) {
+			minecraft.getTextureManager().bindTexture(TEXTURE);
+			GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
+			GlStateManager.enableBlend();
+			GlStateManager.tryBlendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
+			GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
+			drawScaledCustomSizeModalRect(x, y, 0, selected ? 20.0F : 0.0F, 20, 20, width, height, 32, 64);
+			minecraft.fontRenderer.drawString(displayString, x + (int) (width * 1.2F), y + (height - 8) / 2, textColor | MathHelper.ceil(255.0F) << 24);
+		}
 	}
 
 	public void setSelected(boolean selected) {
@@ -44,5 +46,15 @@ public class CallbackCheckbox extends ClickButton {
 
 	public boolean selected() {
 		return selected;
+	}
+
+	@Override
+	public int getCurrentIndex() {
+		return selected() ? 1 : 0;
+	}
+
+	@Override
+	public void setCurrentIndex(int newIndex) {
+		setSelected(newIndex == 1);
 	}
 }